# Fast Cache - Downgrade - reupgrade protection and other improvements