--- a/fast_node.go
+++ b/fast_node.go
@@ -12,7 +12,7 @@ type FastNode struct {
 	key                  []byte
 	versionLastUpdatedAt int64
 	value                []byte
-	// leafHash             []byte // TODO: Look into if this would help with proof stuff.
+	leafHash             []byte
 }
 
 // NewFastNode returns a new fast node from a value and version.
@@ -32,22 +32,32 @@ func DeserializeFastNode(key []byte, buf []byte) (*FastNode, error) {
 	}
 	buf = buf[n:]
 
-	val, _, cause := decodeBytes(buf)
+	val, n, cause := decodeBytes(buf)
 	if cause != nil {
-		return nil, errors.Wrap(cause, "decoding node.value")
+		return nil, errors.Wrap(cause, "decoding fastnode.value")
+	}
+	buf = buf[n:]
+
+	leafHash, _, cause := decodeBytes(buf)
+	if cause != nil {
+		return nil, errors.Wrap(cause, "decoding fastnode.leafHash")
 	}
 
 	fastNode := &FastNode{
 		key:                  key,
 		versionLastUpdatedAt: ver,
 		value:                val,
+		leafHash:        leafHash,
 	}
 
 	return fastNode, nil
 }
 
 func (node *FastNode) encodedSize() int {
-	n := encodeVarintSize(node.versionLastUpdatedAt) + encodeBytesSize(node.value)
+	n :=
+	encodeVarintSize(node.versionLastUpdatedAt) +
+	encodeBytesSize(node.value) +
+	encodeBytesSize(node.leafHash)
 	return n
 }
 
@@ -63,6 +73,10 @@ func (node *FastNode) writeBytes(w io.Writer) error {
 	cause = encodeBytes(w, node.value)
 	if cause != nil {
 		return errors.Wrap(cause, "writing value")
+	}
+	cause = encodeBytes(w, node.leafHash)
+	if cause != nil {
+		return errors.Wrap(cause, "writing leaf hash")
 	}
 	return nil
 }