--- a/immutable_tree.go
+++ b/immutable_tree.go
@@ -1,7 +1,6 @@
 package iavl
 
 import (
-	"bytes"
 	"fmt"
 	"strings"