**#11 - Background**

We would like to add downgrade and re-upgrade protection for the fast cache. Write the last fast node version on each block and check it at the startup. If not set to the most recent version, rewrite fast cache

**Acceptance criteria**
- [x] protection added
- [x] unit tested


**Additionally:**
- [x] optimized non-membership proof
- [x] unit tested fast node
- [x] resolved a bug related to the location of where migration was applied in `LoadVersion` and `LazyLoadVersion` 
- [x] looked into optimizing `GetWithIdndex` and `GetByIndex`
    * the only idea made it worse off so had to revert
- [x] bench tests for `GetWithIndex`, `GetByIndex` and `  


Latest benchmark for non-membership proof:

```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkGetNonMembership$ github.com/cosmos/iavl

goos: linux
goarch: amd64
pkg: github.com/cosmos/iavl
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkGetNonMembership/fast-8         	    1130	   1003260 ns/op	  132431 B/op	    2920 allocs/op
BenchmarkGetNonMembership/regular-8      	     154	   7399486 ns/op	 1292050 B/op	   21474 allocs/op
PASS
ok  	github.com/cosmos/iavl	54.441s
```

Old benchmark from `dev/iavl_data_locality` with modified bench tests:
```
iavl: 
git commit: 
git branch: 
go version go1.17.6 linux/amd64

Init Tree took 78.76 MB
goos: linux
goarch: amd64
pkg: github.com/cosmos/iavl/benchmarks
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkMedium/memdb-100000-100-16-40/query-no-in-tree-guarantee-slow-8         	   93859	     12319 ns/op	     592 B/op	      12 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/query-hits-slow-8                         	   83343	     14889 ns/op	     760 B/op	      15 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/iteration-slow-8                          	       2	 829462950 ns/op	88838692 B/op	 1600081 allocs/op
--- BENCH: BenchmarkMedium/memdb-100000-100-16-40/iteration-slow-8
    bench_test.go:128: completed 100000 iterations
    bench_test.go:128: completed 100000 iterations
    bench_test.go:128: completed 100000 iterations
BenchmarkMedium/memdb-100000-100-16-40/update-8                                  	   10000	    149613 ns/op	   27337 B/op	     335 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/block-8                                   	      78	  16881365 ns/op	 2910816 B/op	   35693 allocs/op
Init Tree took 40.64 MB
BenchmarkMedium/goleveldb-100000-100-16-40/query-no-in-tree-guarantee-slow-8     	   56197	     21425 ns/op	    1566 B/op	      30 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/query-hits-slow-8                     	   43808	     27835 ns/op	    2122 B/op	      39 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/iteration-slow-8                      	       1	1974250600 ns/op	226939520 B/op	 3862227 allocs/op
--- BENCH: BenchmarkMedium/goleveldb-100000-100-16-40/iteration-slow-8
    bench_test.go:128: completed 100000 iterations
BenchmarkMedium/goleveldb-100000-100-16-40/update-8                              	    8377	    211953 ns/op	   43961 B/op	     425 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/block-8                               	      45	  28982847 ns/op	 5162017 B/op	   53449 allocs/op
PASS
ok  	github.com/cosmos/iavl/benchmarks	25.442s
```

Latest benchmark:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkMedium$ github.com/cosmos/iavl/benchmarks

iavl: 
git commit: 
git branch: 
go version go1.17.6 linux/amd64

Init Tree took 114.29 MB
goos: linux
goarch: amd64
pkg: github.com/cosmos/iavl/benchmarks
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkMedium/memdb-100000-100-16-40/query-no-in-tree-guarantee-fast-8         	  596952	      2612 ns/op	     112 B/op	       4 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/query-no-in-tree-guarantee-slow-8         	   96513	     13278 ns/op	     440 B/op	       8 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/query-hits-fast-8                         	  737870	      1686 ns/op	      26 B/op	       0 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/query-hits-slow-8                         	   69254	     19333 ns/op	     605 B/op	      11 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/iteration-fast-8                          	      16	 101840581 ns/op	18400314 B/op	  300000 allocs/op
--- BENCH: BenchmarkMedium/memdb-100000-100-16-40/iteration-fast-8
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
	... [output truncated]
BenchmarkMedium/memdb-100000-100-16-40/iteration-slow-8                          	       2	1075280050 ns/op	88843124 B/op	 1600102 allocs/op
--- BENCH: BenchmarkMedium/memdb-100000-100-16-40/iteration-slow-8
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
BenchmarkMedium/memdb-100000-100-16-40/update-8                                  	    8095	    226336 ns/op	   27376 B/op	     341 allocs/op
BenchmarkMedium/memdb-100000-100-16-40/block-8                                   	      52	  22879244 ns/op	 2946879 B/op	   36431 allocs/op
Init Tree took 66.82 MB
BenchmarkMedium/goleveldb-100000-100-16-40/query-no-in-tree-guarantee-fast-8     	  238384	      6958 ns/op	     814 B/op	      16 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/query-no-in-tree-guarantee-slow-8     	   33271	     34922 ns/op	    1844 B/op	      32 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/query-hits-fast-8                     	  459201	      2207 ns/op	     114 B/op	       2 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/query-hits-slow-8                     	   38526	     32942 ns/op	    1987 B/op	      33 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/iteration-fast-8                      	      10	 109896370 ns/op	29327281 B/op	  522988 allocs/op
--- BENCH: BenchmarkMedium/goleveldb-100000-100-16-40/iteration-fast-8
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
    bench_test.go:143: completed 100000 iterations
	... [output truncated]
BenchmarkMedium/goleveldb-100000-100-16-40/iteration-slow-8                      	       1	2197231400 ns/op	236026680 B/op	 3998799 allocs/op
--- BENCH: BenchmarkMedium/goleveldb-100000-100-16-40/iteration-slow-8
    bench_test.go:143: completed 100000 iterations
BenchmarkMedium/goleveldb-100000-100-16-40/update-8                              	    7573	    281499 ns/op	   43491 B/op	     432 allocs/op
BenchmarkMedium/goleveldb-100000-100-16-40/block-8                               	      30	  49443073 ns/op	 6370082 B/op	   66312 allocs/op
PASS
ok  	github.com/cosmos/iavl/benchmarks	49.589s
```