@akhtariev 2022-01-29
It seems that GetwithIndex and GetByIndex have actually become worse off:
```
// OLD Implementation

Benchmark_GetWithIndex/fast-8         	   55580	     18788 ns/op	     764 B/op	      15 allocs/op
Benchmark_GetWithIndex/regular-8      	   78734	     17354 ns/op	     720 B/op	      15 allocs/op
PASS
ok  	github.com/cosmos/iavl	5.688s


Benchmark_GetByIndex/fast-8         	   66823	     17711 ns/op	     828 B/op	      15 allocs/op
Benchmark_GetByIndex/regular-8      	   71974	     17146 ns/op	     822 B/op	      15 allocs/op
PASS
ok  	github.com/cosmos/iavl	6.172s


// NEW Implementation

Benchmark_GetWithIndex/fast-8         	      44	  23212700 ns/op	 3865367 B/op	   96616 allocs/op
Benchmark_GetWithIndex/regular-8      	   85010	     14220 ns/op	     584 B/op	      11 allocs/op
PASS
ok  	github.com/cosmos/iavl	7.435s

Benchmark_GetByIndex/fast-8         	     100	  20810321 ns/op	 3821985 B/op	   95532 allocs/op
Benchmark_GetByIndex/regular-8      	   72451	     17056 ns/op	     828 B/op	      15 allocs/op
PASS
ok  	github.com/cosmos/iavl	7.200s
```

As a result, using fast iterations to GetWithIndex and GetByIndex might not be the most optimal solution. Will try reverting the changes back to where they were...