@akhtariev 2022-01-31
@ValarDragon addressed all comments so far. The most recent updates:

- commented out `GetNonMembershipProof` new implementation for now
- addressed the bug about appending version and unit tested
- added more unit tests in mutable_tree_test.go for re-upgrade protection
- renamed some methods for clarity and added comments

Please take a look when you have time